requirements.txt
flask
Flask-SQLAlchemy
flask-restful
python-dotenv